spike: simplify test harness
