+++ crates/client/test-utils/src/node.rs
@@ -1,23 +1,12 @@
 //! Local node setup with Base Sepolia chainspec
 
-use std::{
-    any::Any,
-    fmt,
-    net::SocketAddr,
-    sync::{Arc, Mutex},
-};
+use std::{any::Any, fmt, net::SocketAddr, sync::Arc};
 
 use alloy_genesis::Genesis;
 use alloy_provider::RootProvider;
 use alloy_rpc_client::RpcClient;
-use base_flashtypes::Flashblock;
-use base_reth_flashblocks::{
-    EthApiExt, EthApiOverrideServer, EthPubSub, EthPubSubApiServer, FlashblocksReceiver,
-    FlashblocksState,
-};
 use eyre::Result;
 use futures_util::Future;
-use once_cell::sync::OnceCell;
 use op_alloy_network::Optimism;
 use reth::{
     api::{FullNodeTypesAdapter, NodeTypesWithDBAdapter},
@@ -34,23 +23,18 @@ use reth_db::{
     test_utils::{ERROR_DB_CREATION, TempDatabase, tempdir_path},
 };
 use reth_e2e_test_utils::{Adapter, TmpDB};
-use reth_exex::ExExEvent;
 use reth_node_core::{
     args::DatadirArgs,
     dirs::{DataDirPath, MaybePlatformPath},
 };
 use reth_optimism_chainspec::OpChainSpec;
 use reth_optimism_node::{OpNode, args::RollupArgs};
-use reth_provider::{CanonStateSubscriptions, providers::BlockchainProvider};
-use tokio::sync::{mpsc, oneshot};
-use tokio_stream::StreamExt;
+use reth_provider::providers::BlockchainProvider;
 
 use crate::engine::EngineApi;
 
 /// Convenience alias for the local blockchain provider type.
 pub type LocalNodeProvider = BlockchainProvider<NodeTypesWithDBAdapter<OpNode, TmpDB>>;
-/// Convenience alias for the Flashblocks state backing the local node.
-pub type LocalFlashblocksState = FlashblocksState<LocalNodeProvider>;
 
 /// Handle to a launched local node along with the resources required to keep it alive.
 pub struct LocalNode {
@@ -70,155 +54,6 @@ impl fmt::Debug for LocalNode {
             .field("ws_api_addr", &self.ws_api_addr)
             .field("engine_ipc_path", &self.engine_ipc_path)
             .finish_non_exhaustive()
-    }
-}
-
-/// Components that allow tests to interact with the Flashblocks worker tasks.
-#[derive(Clone)]
-pub struct FlashblocksParts {
-    sender: mpsc::Sender<(Flashblock, oneshot::Sender<()>)>,
-    state: Arc<LocalFlashblocksState>,
-}
-
-impl fmt::Debug for FlashblocksParts {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("FlashblocksParts").finish_non_exhaustive()
-    }
-}
-
-impl FlashblocksParts {
-    /// Clone the shared [`FlashblocksState`] handle.
-    pub fn state(&self) -> Arc<LocalFlashblocksState> {
-        self.state.clone()
-    }
-
-    /// Send a flashblock to the background processor and wait until it is handled.
-    pub async fn send(&self, flashblock: Flashblock) -> Result<()> {
-        let (tx, rx) = oneshot::channel();
-        self.sender.send((flashblock, tx)).await.map_err(|err| eyre::eyre!(err))?;
-        rx.await.map_err(|err| eyre::eyre!(err))?;
-        Ok(())
-    }
-}
-
-#[derive(Clone)]
-struct FlashblocksNodeExtensions {
-    inner: Arc<FlashblocksNodeExtensionsInner>,
-}
-
-struct FlashblocksNodeExtensionsInner {
-    sender: mpsc::Sender<(Flashblock, oneshot::Sender<()>)>,
-    #[allow(clippy::type_complexity)]
-    receiver: Arc<Mutex<Option<mpsc::Receiver<(Flashblock, oneshot::Sender<()>)>>>>,
-    fb_cell: Arc<OnceCell<Arc<LocalFlashblocksState>>>,
-    process_canonical: bool,
-}
-
-impl FlashblocksNodeExtensions {
-    fn new(process_canonical: bool) -> Self {
-        let (sender, receiver) = mpsc::channel::<(Flashblock, oneshot::Sender<()>)>(100);
-        let inner = FlashblocksNodeExtensionsInner {
-            sender,
-            receiver: Arc::new(Mutex::new(Some(receiver))),
-            fb_cell: Arc::new(OnceCell::new()),
-            process_canonical,
-        };
-        Self { inner: Arc::new(inner) }
-    }
-
-    fn apply(&self, builder: OpBuilder) -> OpBuilder {
-        let fb_cell = self.inner.fb_cell.clone();
-        let receiver = self.inner.receiver.clone();
-        let process_canonical = self.inner.process_canonical;
-
-        let fb_cell_for_exex = fb_cell.clone();
-
-        builder
-            .install_exex("flashblocks-canon", move |mut ctx| {
-                let fb_cell = fb_cell_for_exex.clone();
-                let process_canonical = process_canonical;
-                async move {
-                    let provider = ctx.provider().clone();
-                    let fb = init_flashblocks_state(&fb_cell, &provider);
-                    Ok(async move {
-                        while let Some(note) = ctx.notifications.try_next().await? {
-                            if let Some(committed) = note.committed_chain() {
-                                let hash = committed.tip().num_hash();
-                                if process_canonical {
-                                    // Many suites drive canonical updates manually to reproduce race conditions, so
-                                    // allowing this to be disabled keeps canonical replay deterministic.
-                                    let chain = Arc::unwrap_or_clone(committed);
-                                    for (_, block) in chain.into_blocks() {
-                                        fb.on_canonical_block_received(block);
-                                    }
-                                }
-                                let _ = ctx.events.send(ExExEvent::FinishedHeight(hash));
-                            }
-                        }
-                        Ok(())
-                    })
-                }
-            })
-            .extend_rpc_modules(move |ctx| {
-                let fb_cell = fb_cell.clone();
-                let provider = ctx.provider().clone();
-                let fb = init_flashblocks_state(&fb_cell, &provider);
-
-                let mut canon_stream = tokio_stream::wrappers::BroadcastStream::new(
-                    ctx.provider().subscribe_to_canonical_state(),
-                );
-                tokio::spawn(async move {
-                    use tokio_stream::StreamExt;
-                    while let Some(Ok(notification)) = canon_stream.next().await {
-                        provider.canonical_in_memory_state().notify_canon_state(notification);
-                    }
-                });
-                let api_ext = EthApiExt::new(
-                    ctx.registry.eth_api().clone(),
-                    ctx.registry.eth_handlers().filter.clone(),
-                    fb.clone(),
-                );
-                ctx.modules.replace_configured(api_ext.into_rpc())?;
-
-                // Register eth_subscribe subscription endpoint for flashblocks
-                // Uses replace_configured since eth_subscribe already exists from reth's standard module
-                // Pass eth_api to enable proxying standard subscription types to reth's implementation
-                let eth_pubsub = EthPubSub::new(ctx.registry.eth_api().clone(), fb.clone());
-                ctx.modules.replace_configured(eth_pubsub.into_rpc())?;
-
-                let fb_for_task = fb.clone();
-                let mut receiver = receiver
-                    .lock()
-                    .expect("flashblock receiver mutex poisoned")
-                    .take()
-                    .expect("flashblock receiver should only be initialized once");
-                tokio::spawn(async move {
-                    while let Some((payload, tx)) = receiver.recv().await {
-                        fb_for_task.on_flashblock_received(payload);
-                        let _ = tx.send(());
-                    }
-                });
-
-                Ok(())
-            })
-    }
-
-    fn wrap_launcher<L, LRet>(&self, launcher: L) -> impl FnOnce(OpBuilder) -> LRet
-    where
-        L: FnOnce(OpBuilder) -> LRet,
-    {
-        let extensions = self.clone();
-        move |builder| {
-            let builder = extensions.apply(builder);
-            launcher(builder)
-        }
-    }
-
-    fn parts(&self) -> Result<FlashblocksParts> {
-        let state = self.inner.fb_cell.get().ok_or_else(|| {
-            eyre::eyre!("FlashblocksState should be initialized during node launch")
-        })?;
-        Ok(FlashblocksParts { sender: self.inner.sender.clone(), state: state.clone() })
     }
 }
 
@@ -373,95 +208,4 @@ where
         _node: Box::new(node_handle),
         _task_manager: tasks,
     })
-}
-
-fn init_flashblocks_state(
-    cell: &Arc<OnceCell<Arc<LocalFlashblocksState>>>,
-    provider: &LocalNodeProvider,
-) -> Arc<LocalFlashblocksState> {
-    cell.get_or_init(|| {
-        let fb = Arc::new(FlashblocksState::new(provider.clone(), 5));
-        fb.start();
-        fb
-    })
-    .clone()
-}
-
-/// Local node wrapper that exposes helpers specific to Flashblocks tests.
-pub struct FlashblocksLocalNode {
-    node: LocalNode,
-    parts: FlashblocksParts,
-}
-
-impl fmt::Debug for FlashblocksLocalNode {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("FlashblocksLocalNode")
-            .field("node", &self.node)
-            .field("parts", &self.parts)
-            .finish()
-    }
-}
-
-impl FlashblocksLocalNode {
-    /// Launch a flashblocks-enabled node using the default launcher.
-    pub async fn new() -> Result<Self> {
-        Self::with_launcher(default_launcher).await
-    }
-
-    /// Builds a flashblocks-enabled node with canonical block streaming disabled so tests can call
-    /// `FlashblocksState::on_canonical_block_received` at precise points.
-    pub async fn manual_canonical() -> Result<Self> {
-        Self::with_manual_canonical_launcher(default_launcher).await
-    }
-
-    /// Launch a flashblocks-enabled node with a custom launcher and canonical processing enabled.
-    pub async fn with_launcher<L, LRet>(launcher: L) -> Result<Self>
-    where
-        L: FnOnce(OpBuilder) -> LRet,
-        LRet: Future<Output = eyre::Result<NodeHandle<Adapter<OpNode>, OpAddOns>>>,
-    {
-        Self::with_launcher_inner(launcher, true).await
-    }
-
-    /// Same as [`Self::with_launcher`] but leaves canonical processing to the caller.
-    pub async fn with_manual_canonical_launcher<L, LRet>(launcher: L) -> Result<Self>
-    where
-        L: FnOnce(OpBuilder) -> LRet,
-        LRet: Future<Output = eyre::Result<NodeHandle<Adapter<OpNode>, OpAddOns>>>,
-    {
-        Self::with_launcher_inner(launcher, false).await
-    }
-
-    async fn with_launcher_inner<L, LRet>(launcher: L, process_canonical: bool) -> Result<Self>
-    where
-        L: FnOnce(OpBuilder) -> LRet,
-        LRet: Future<Output = eyre::Result<NodeHandle<Adapter<OpNode>, OpAddOns>>>,
-    {
-        let extensions = FlashblocksNodeExtensions::new(process_canonical);
-        let wrapped_launcher = extensions.wrap_launcher(launcher);
-        let node = LocalNode::new(wrapped_launcher).await?;
-
-        let parts = extensions.parts()?;
-        Ok(Self { node, parts })
-    }
-
-    /// Access the shared Flashblocks state for assertions or manual driving.
-    pub fn flashblocks_state(&self) -> Arc<LocalFlashblocksState> {
-        self.parts.state()
-    }
-
-    /// Send a flashblock through the background processor and await completion.
-    pub async fn send_flashblock(&self, flashblock: Flashblock) -> Result<()> {
-        self.parts.send(flashblock).await
-    }
-
-    /// Split the wrapper into the underlying node plus flashblocks parts.
-    pub fn into_parts(self) -> (LocalNode, FlashblocksParts) {
-        (self.node, self.parts)
-    }
-
-    /// Borrow the underlying [`LocalNode`].
-    pub fn as_node(&self) -> &LocalNode {
-        &self.node
-    }
 }
